+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 2.10.2 (pending)
 
 * added Gitea support for IndexModules
+* changed system property `lavender.bitbucket.wirelog` to `lavender.wirelog`
 * dumped explicitly rejecting pustefix.relative
 * updated debian dependencies from Java 8 to 11
 * no parent pom update, because it would compile for java 16, which is not yet an option